Move TemplateFixture to its own file
